requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
multidict==6.1.0
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.59.6
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
pytesseract==0.3.13
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
